spring boot testing fix tests
